Update LAB_03b-Manage_Azure_Resources_by_Using_ARM_Templates.md
I was practicing this lab and found that it doesn't provide any drop down or any default value for Location and one needs to enter that. I tried using the same as for resource group which was (US) East US, however the validation failed saying that this location is not available and the available locations are "southeastasia, eastus2" etc. I copied this as it is and when I validated the same. this time the deployment validation went through as well as completed successfully.

Hence I am suggesting this change.